Split value conversion into a separate module
